Move debuginfo extraction to elfwriter package
The debuginfo package has dependencies to Linux only things which makes
it hard to test extraction on osx. This is primarily to make development
easier.
